Renaming boom/alert.py to boom/object.rs and boom/get_alerts.py to boom/alert.rs. Renaming BoomGetAlertsHandler to BoomAlertsHandler.
