+++ skyportal/handlers/api/boom/__init__.py
@@ -1,5 +1,5 @@
-from .alert import BoomObjectHandler
+from .alert import BoomAlertsHandler
 from .filter import BoomFilterHandler
 from .filter_modules import BoomFilterModulesHandler
-from .get_alerts import BoomGetAlertsHandler
+from .object import BoomObjectHandler
 from .run_filter import BoomRunFilterHandler
